# webapp/src/main/resources/openapi.yaml
openapi: 3.0.1
info:
  description: Mojito API Documentation
  title: Mojito API
  version: 0.111-SNAPSHOT
servers:
- description: Generated server url
  url: http://localhost:8080
paths:
...
  /api/commits:
    get:
      operationId: getCommits
      parameters:
      - in: query
        name: repositoryId
        required: true
        schema:
          type: integer
          format: int64
      - in: query
        name: commitNames
        required: false
        schema:
          type: array
          items:
            type: string
      - in: query
        name: pushRunName
        required: false
        schema:
          type: string
      - in: query
        name: pullRunName
        required: false
        schema:
          type: string
      - in: query
        name: hasPushRun
        required: false
        schema:
          type: boolean
      - in: query
        name: hasPullRun
        required: false
        schema:
          type: boolean
      - description: Zero-based page index (0..N)
        in: query
        name: page
        required: false
        schema:
          type: integer
          default: 0
          minimum: 0
      - description: The size of the page to be returned
        in: query
        name: size
        required: false
        schema:
          type: integer
          default: 10
          minimum: 1
      - description: "Sorting criteria in the format: property,(asc|desc). Default\
          \ sort order is ascending. Multiple sort criteria are supported."
        in: query
        name: sort
        required: false
        schema:
          type: array
          default:
          - "id,DESC"
          items:
            type: string
      responses:
        "200":
          content:
            application/hal+json:
              schema:
                $ref: "#/components/schemas/PageCommit_Commit"
          description: OK
      tags:
      - commit-ws
    post:
      operationId: createCommit
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommitBody"
        required: true
      responses:
        "200":
          content:
            application/hal+json:
              schema:
                $ref: "#/components/schemas/Commit_Commit"
          description: OK
      tags:
      - commit-ws
  /api/commits/detailed:
    get:
      operationId: getCommitsDetailed
      parameters:
      - in: query
        name: repositoryId
        required: true
        schema:
          type: integer
          format: int64
      - in: query
        name: commitNames
        required: false
        schema:
          type: array
          items:
            type: string
      - in: query
        name: pushRunName
        required: false
        schema:
          type: string
      - in: query
        name: pullRunName
        required: false
        schema:
          type: string
      - in: query
        name: hasPushRun
        required: false
        schema:
          type: boolean
      - in: query
        name: hasPullRun
        required: false
        schema:
          type: boolean
      - description: Zero-based page index (0..N)
        in: query
        name: page
        required: false
        schema:
          type: integer
          default: 0
          minimum: 0
      - description: The size of the page to be returned
        in: query
        name: size
        required: false
        schema:
          type: integer
          default: 10
          minimum: 1
      - description: "Sorting criteria in the format: property,(asc|desc). Default\
          \ sort order is ascending. Multiple sort criteria are supported."
        in: query
        name: sort
        required: false
        schema:
          type: array
          default:
          - "id,DESC"
          items:
            type: string
      responses:
        "200":
          content:
            application/hal+json:
              schema:
                $ref: "#/components/schemas/PageCommit_CommitDetailed"
          description: OK
      tags:
      - commit-ws
  /api/commits/lastPullRun/:
    post:
      operationId: getLastPullRun
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommitListWithRepositoryIdBody"
        required: true
      responses:
        "200":
          content:
            application/hal+json:
              schema:
                $ref: "#/components/schemas/PullRun_CommitDetailed"
          description: OK
      tags:
      - commit-ws
  /api/commits/lastPulled/:
    post:
      operationId: getLastPulledCommit
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommitListWithRepositoryIdBody"
        required: true
      responses:
        "200":
          content:
            application/hal+json:
              schema:
                $ref: "#/components/schemas/Commit_Commit"
          description: OK
      tags:
      - commit-ws
  /api/commits/lastPushRun/:
    post:
      operationId: getLastPushRun
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommitListWithRepositoryIdBody"
        required: true
      responses:
        "200":
          content:
            application/hal+json:
              schema:
                $ref: "#/components/schemas/PushRun_CommitDetailed"
          description: OK
      tags:
      - commit-ws
  /api/commits/lastPushed/:
    post:
      operationId: getLastPushedCommit
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommitListWithRepositoryIdBody"
        required: true
      responses:
        "200":
          content:
            application/hal+json:
              schema:
                $ref: "#/components/schemas/Commit_Commit"
          description: OK
      tags:
      - commit-ws
  /api/commits/pullRun:
    post:
      operationId: associateCommitToPullRun
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommitToPullRunBody"
        required: true
      responses:
        "200":
          description: OK
      tags:
      - commit-ws
  /api/commits/pushRun:
    post:
      operationId: associateCommitToPushRun
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommitToPushRunBody"
        required: true
      responses:
        "200":
          description: OK
      tags:
      - commit-ws
...
components:
  schemas:
...
    CommitBody:
      type: object
      properties:
        appendBranchTextUnitsId:
          type: string
        authorEmail:
          type: string
        authorName:
          type: string
        commitName:
          type: string
        repositoryId:
          type: integer
          format: int64
        sourceCreationDate:
          type: integer
          format: int64
          example: 1715699917000
        sourceCreationDate2:
          type: string
    CommitListWithRepositoryIdBody:
      type: object
      properties:
        commitNames:
          type: array
          items:
            type: string
        repositoryId:
          type: integer
          format: int64
    CommitToPullRunBody:
      type: object
      properties:
        commitName:
          type: string
        pullRunName:
          type: string
        repositoryId:
          type: integer
          format: int64
    CommitToPullRun_CommitDetailed:
      type: object
      properties:
        createdDate:
          type: integer
          format: int64
          example: 1715699917000
        id:
          type: integer
          format: int64
        pullRun:
          $ref: "#/components/schemas/PullRun_CommitDetailed"
    CommitToPushRunBody:
      type: object
      properties:
        commitName:
          type: string
        pushRunName:
          type: string
        repositoryId:
          type: integer
          format: int64
    CommitToPushRun_CommitDetailed:
      type: object
      properties:
        createdDate:
          type: integer
          format: int64
          example: 1715699917000
        id:
          type: integer
          format: int64
        pushRun:
          $ref: "#/components/schemas/PushRun_CommitDetailed"
    Commit_Commit:
      type: object
      properties:
        authorEmail:
          type: string
        authorName:
          type: string
        createdDate:
          type: integer
          format: int64
          example: 1715699917000
        id:
          type: integer
          format: int64
        lastModifiedDate:
          type: integer
          format: int64
          example: 1715699917000
        name:
          type: string
        repository_id:
          type: integer
          format: int64
        sourceCreationDate:
          type: integer
          format: int64
          example: 1715699917000
    Commit_CommitDetailed:
      type: object
      properties:
        authorEmail:
          type: string
        authorName:
          type: string
        commitToPullRun:
          $ref: "#/components/schemas/CommitToPullRun_CommitDetailed"
        commitToPushRun:
          $ref: "#/components/schemas/CommitToPushRun_CommitDetailed"
        createdDate:
          type: integer
          format: int64
          example: 1715699917000
        id:
          type: integer
          format: int64
        lastModifiedDate:
          type: integer
          format: int64
          example: 1715699917000
        name:
          type: string
        repository_id:
          type: integer
          format: int64
        sourceCreationDate:
          type: integer
          format: int64
          example: 1715699917000
...
    PageCommit_Commit:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/Commit_Commit"
        empty:
          type: boolean
        first:
          type: boolean
        last:
          type: boolean
        number:
          type: integer
          format: int32
        numberOfElements:
          type: integer
          format: int32
        pageable:
          $ref: "#/components/schemas/PageableObject_Commit"
        size:
          type: integer
          format: int32
        sort:
          $ref: "#/components/schemas/SortObject"
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
    PageCommit_CommitDetailed:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/Commit_CommitDetailed"
        empty:
          type: boolean
        first:
          type: boolean
        last:
          type: boolean
        number:
          type: integer
          format: int32
        numberOfElements:
          type: integer
          format: int32
        pageable:
          $ref: "#/components/schemas/PageableObject_CommitDetailed"
        size:
          type: integer
          format: int32
        sort:
          $ref: "#/components/schemas/SortObject"
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
...
    PageableObject_Commit:
      type: object
    PageableObject_CommitDetailed:
      type: object
...
    PullRun_CommitDetailed:
      type: object
      properties:
        createdDate:
          type: integer
          format: int64
          example: 1715699917000
        id:
          type: integer
          format: int64
        name:
          type: string
    PushRun_CommitDetailed:
      type: object
      properties:
        createdDate:
          type: integer
          format: int64
          example: 1715699917000
        id:
          type: integer
          format: int64
        name:
          type: string
...
    SortObject:
      type: object
      properties:
        empty:
          type: boolean
        sorted:
          type: boolean
        unsorted:
          type: boolean
...
  securitySchemes:
    cookieAuth:
      in: cookie
      name: JSESSIONID
      type: apiKey